+++ .gitignore
@@ -3,6 +3,10 @@ __pycache__/
 *.py[cod]
 *$py.class
 
+
+#Database
+worldstats.db
+
 # C extensions
 *.so
