--- responses/Communism.py
@@ -1,36 +1,0 @@
-arr_communism = [
-    "Avante, camarada",
-    "Saudades da Venezuela",
-    "Das Kapital",
-    "Cala-te, seu comuna",
-    "Bem-vindo de volta, comuna",
-    "Vamos ler o manifesto comunista",
-    "Vamos ouvir a Carvalhesa",
-    "Tu, eu, e a conquista dos meios de produção?",
-    "Proletários de todos os países, uni-vos!",
-    "Tá calado, comuna",
-    "Stalin matou foi pouco",
-    "Ide plantar batatas, seu comuna",
-    "Agora não, camarada",
-    "Malditos comunas",
-    "Alerta comuna",
-    "Stalin era muito daddy",
-    "Boas, camarada, vamos unir os meios de reprodução?",
-    "Vamos produzir prole para a nossa Pátria Mãe?",
-    "És um meio de produção? Quero conquistar-te",
-    "If your legs were wealth I'd spread them equally",
-    "Roses are red, Communism is great, How about we both go on a date?",
-    "Hey girl, do you want to Fidel my Castro?",
-    "I want to put my 13 inches of HARD labour in your gulag",
-    "És tão linda que olhei para ti e as minhas bochechas ficaram da cor do manifesto comunista",
-    "Ruin me like Stalin ruined everyone’s lives",
-    "Are you the means of production? Because I can’t stop thinking about how we could create something beautiful together",
-    "Coletiviza-mos",
-    "Não sou Karl Marx, mas o meu amor por ti dava um manifesto",
-    "Só não te dou o mundo porque ele pertence à classe trabalhadora",
-    "A única propriedade privada que quero é a do teu coração",
-    "Roubaste o meu coração tal como os burgueses roubam a classe trabalhadora",
-    "Vem ser o martelo da minha foice",
-    "Chama-me comunista porque por mim partilhávamos cama",
-    "Mais red flags que a URSS",    
-]
