+++ responses/Custom.py
@@ -1,0 +1,183 @@
+arr_communism = [
+    "Avante, camarada",
+    "Saudades da Venezuela",
+    "Das Kapital",
+    "Cala-te, seu comuna",
+    "Bem-vindo de volta, comuna",
+    "Vamos ler o manifesto comunista",
+    "Vamos ouvir a Carvalhesa",
+    "Tu, eu, e a conquista dos meios de produção?",
+    "Proletários de todos os países, uni-vos!",
+    "Tá calado, comuna",
+    "Stalin matou foi pouco",
+    "Ide plantar batatas, seu comuna",
+    "Agora não, camarada",
+    "Malditos comunas",
+    "Alerta comuna",
+    "Stalin era muito daddy",
+    "Boas, camarada, vamos unir os meios de reprodução?",
+    "Vamos produzir prole para a nossa Pátria Mãe?",
+    "És um meio de produção? Quero conquistar-te",
+    "If your legs were wealth I'd spread them equally",
+    "Roses are red, Communism is great, How about we both go on a date?",
+    "Hey girl, do you want to Fidel my Castro?",
+    "I want to put my 13 inches of HARD labour in your gulag",
+    "És tão linda que olhei para ti e as minhas bochechas ficaram da cor do manifesto comunista",
+    "Ruin me like Stalin ruined everyone’s lives",
+    "Are you the means of production? Because I can’t stop thinking about how we could create something beautiful together",
+    "Coletiviza-mos",
+    "Não sou Karl Marx, mas o meu amor por ti dava um manifesto",
+    "Só não te dou o mundo porque ele pertence à classe trabalhadora",
+    "A única propriedade privada que quero é a do teu coração",
+    "Roubaste o meu coração tal como os burgueses roubam a classe trabalhadora",
+    "Vem ser o martelo da minha foice",
+    "Chama-me comunista porque por mim partilhávamos cama",
+    "Mais red flags que a URSS",    
+]
+
+arr_euskara = [
+    "Egun on",
+    "Euskadi Ta Askatasuna",
+    "Txerria txerria txerria",
+    "Frossos Tripodea",
+    "Kabroi!",
+    "Zozoa!",   
+]
+
+arr_french = [
+    "Tu es lent comme un escargot!",
+    "Tu as deux mains gauches!",
+    "Tu es un vrai moulin à paroles.",
+    "On ne t’a pas sonné!",
+    "T’as pris la grosse tête!",
+    "Tu as un petit vélo dans la tête.",
+    "T’es haut comme trois pommes.",
+    "Tu racontes des salades.",
+    "T’es pas une lumière!",
+    "T’as dormi dans une éolienne?",
+    "T’as une tête à faire peur aux enfants!",
+    "C’est pas demain que tu seras Einstein.",
+    "Pain au chocolat",
+    "Omellete du fromage",
+    "Oui oui baguette",
+    "Alors on danse",
+    "C'est l'monde à l'envers",
+    "Je vois la vie en rose",
+    "Je te laisserai des mots",
+    "Merde",
+    "Je t'aime",
+    "C'est la vie",
+    "Casse toi",
+    "Puis-je vous offrir un verre?",
+    "J’ai eu un coup de foudre",
+    "Sans toi, je ne suis rien",
+    "Je m'en fous",
+    "C'est dommage",
+    "Je ne sais pas",
+    "Mettre les couilles sur la table",
+    "C'est comme pisser dans un violon",
+    "Mets la moi dans le cul",
+]
+
+arr_german = [
+    "Ich lerne Deutsch für dich <3",
+    "Es ist Mamma",
+    "Heil JardasBot",
+    "Du bist mein Liebling",
+    "Scheiße",
+    "Scheiß drauf, wir Ball",
+    "gespannt",
+    "Ich, deine Mutter",
+    "Das war's und hör auf, dumm zu sein",
+    "90 Tage Verlobte heute?",
+    "Bulbasaur ist das beste Pokémon",
+    "Deine Locken sind heute sehr schön",
+    "Englisch ist mein Lieblingsfach, Lieblingsfach, Lieblingsfach",
+    "Bratwurst",
+    "Sehr gut, Mamma",
+    "Kann ich meine Pommes in Ihren McFlurry dippen?",
+    "Schwester",
+    "Lass uns Tripa gehen",
+    "Lass uns Emo-Musik hören",
+    "Lucy und Mia sind die besten Katzen",
+    "Ich glaub mein Schwein pfeift",
+    "Das ist mir Wurst",
+    "Alles in Butter",
+    "Einen Vogel haben",
+    "Das Leben ist kein Ponyhof",
+    "Schönen Feierabend",
+    "Hast Du Bock?",
+    "Voll Geil",
+    "Heil, Chancellora",
+    "Ich tu dir weh",
+    "Schauen wir uns Game of Thrones an",
+]
+
+arr_japanese = [
+    "ドラマー",
+    "Ichi, ni, san, shi, go, roku, nana, hachi, kyu, jyuu",
+    "Arigato, Renato-kun",
+    "Ā, akogare no pokemon masutā ni",
+    "Konnichiwa",
+    "Sayōnara",
+    "Duolingo no jikan",
+    "Bonifashio wa bakadayo",
+    "バカ",
+    "鶏肉",
+    "Nani?!",
+    "マジェ",
+    "同意します",
+    "Notice me senpai",
+    "ガンダ クロム",
+    "Moshi moshi",
+    "Damare",
+    "Sonikku za Hejjihoggu",
+    "Koshinuke",
+    "大きな腕",
+    "Kuso",
+    "KFC ni ikimashou ka?",
+    "仕事に戻る",
+    "おっぱい",
+    "Jimu ni modoru jikanda",
+    "Baiosenōshisu",
+    "Poruko poruko poruko",
+    "Bamosu jogā monoporio do Sonikku?",
+    "Kinshi no jikandesu",
+    "Afutā 8 aisukurīmu",
+    "Pika pika",
+]
+
+arr_latin = [
+    "Carpe diem quam minimum credula postero",
+    "Tua mater latior quam Rubicon est",
+    "Quid quid latine dictum sit, altum videtur",
+    "In vino veritas",
+    "Cave Canem",
+    "Dictum factum.",
+    "Dum vita est, spes est",
+    "Sapientia potentia est",
+    "Pax vobiscum",
+    "Suus 'tempus ad Duolingo lectionem",
+    "Accusare nemo se debet nisi coram Deo",
+    "Veni, vidi, vici",
+    "Sic parvis magna",
+    "Impetu vobiscum",
+    "Caecilius est in horto!",
+    "Per aspera ad astra",
+    "Audentes Fortuna iuvat",
+    "Alea iacta est",
+    "Et tu, Kiké?",
+    "Mea culpa",
+    "Cogito, ergo sum",
+    "Pedicabo ego vos et irrumabo",
+    "Loqui melius an mori?",
+    "Ave, Biocenose, morituri te salutant",
+    "Pro rata",
+    "Memento mori",
+    "Omnia vincit amor",
+    "Ceterum censeo Carthaginem delendam esse",
+    "Tempora mutantur et nos mutamur in illis",
+    "Forsan et haec olim meminisse iuvabit",
+    "Ignoramus et ignorabimus",
+    "A far l'amore comincia tu",
+]
