+++ responses/DarkJokes.py
@@ -3,7 +3,6 @@ arr_darkjokes = [
     "Dark jokes are like Make-A-Wish kids, they never get old",
     "Where does a suicide bomber go when he dies? Everywhere",
     "What's the difference between Adolf Hitler and Usain Bolt? One can actually finish a race.",
-    "Statistically, 9 out of 10 people enjoy gang rape",
     "Gender is a lot like the Twin Towers, There used to be 2 of them but now it's a sensitive subject",
     "My German girlfriend always yells her age when we have sex",
     "Do you know the difference in peanut butter and jam? Well, I can't peanut butter my dick in your ass",
