--- responses/Euskara.py
@@ -1,8 +1,0 @@
-arr_euskara = [
-    "Egun on",
-    "Euskadi Ta Askatasuna",
-    "Txerria txerria txerria",
-    "Frossos Tripodea",
-    "Kabroi!",
-    "Zozoa!",   
-]
