--- responses/French.py
@@ -1,37 +1,0 @@
-arr_french = [
-    "Tu es lent comme un escargot!",
-    "Tu as deux mains gauches!",
-    "Tu es un vrai moulin à paroles.",
-    "On ne t’a pas sonné!",
-    "T’as pris la grosse tête!",
-    "Tu as un petit vélo dans la tête.",
-    "T’es haut comme trois pommes.",
-    "Tu racontes des salades.",
-    "T’es pas une lumière!",
-    "T’as dormi dans une éolienne?",
-    "T’as une tête à faire peur aux enfants!",
-    "C’est pas demain que tu seras Einstein.",
-    "Pain au chocolat",
-    "Omellete du fromage",
-    "Oui oui baguette",
-    "Alors on danse",
-    "C'est l'monde à l'envers",
-    "Je vois la vie en rose",
-    "Je te laisserai des mots",
-    "Merde",
-    "Je t'aime",
-    "C'est la vie",
-    "Casse toi",
-    "Puis-je vous offrir un verre?",
-    "J’ai eu un coup de foudre",
-    "Sans toi, je ne suis rien",
-    "Je m'en fous",
-    "C'est dommage",
-    "Je ne sais pas",
-    "Mettre les couilles sur la table",
-    "C'est comme pisser dans un violon",
-    "Mets la moi dans le cul",
-    ]
-
-
-
