--- responses/German.py
@@ -1,33 +1,0 @@
-arr_german = [
-    "Ich lerne Deutsch für dich <3",
-    "Es ist Mamma",
-    "Heil JardasBot",
-    "Du bist mein Liebling",
-    "Scheiße",
-    "Scheiß drauf, wir Ball",
-    "gespannt",
-    "Ich, deine Mutter",
-    "Das war's und hör auf, dumm zu sein",
-    "90 Tage Verlobte heute?",
-    "Bulbasaur ist das beste Pokémon",
-    "Deine Locken sind heute sehr schön",
-    "Englisch ist mein Lieblingsfach, Lieblingsfach, Lieblingsfach",
-    "Bratwurst",
-    "Sehr gut, Mamma",
-    "Kann ich meine Pommes in Ihren McFlurry dippen?",
-    "Schwester",
-    "Lass uns Tripa gehen",
-    "Lass uns Emo-Musik hören",
-    "Lucy und Mia sind die besten Katzen",
-    "Ich glaub mein Schwein pfeift",
-    "Das ist mir Wurst",
-    "Alles in Butter",
-    "Einen Vogel haben",
-    "Das Leben ist kein Ponyhof",
-    "Schönen Feierabend",
-    "Hast Du Bock?",
-    "Voll Geil",
-    "Heil, Chancellora",
-    "Ich tu dir weh",
-    "Schauen wir uns Game of Thrones an",
-]
