--- responses/Japanese.py
@@ -1,33 +1,0 @@
-arr_japanese = [
-    "ドラマー",
-    "Ichi, ni, san, shi, go, roku, nana, hachi, kyu, jyuu",
-    "Arigato, Renato-kun",
-    "Ā, akogare no pokemon masutā ni",
-    "Konnichiwa",
-    "Sayōnara",
-    "Duolingo no jikan",
-    "Bonifashio wa bakadayo",
-    "バカ",
-    "鶏肉",
-    "Nani?!",
-    "マジェ",
-    "同意します",
-    "Notice me senpai",
-    "ガンダ クロム",
-    "Moshi moshi",
-    "Damare",
-    "Sonikku za Hejjihoggu",
-    "Koshinuke",
-    "大きな腕",
-    "Kuso",
-    "KFC ni ikimashou ka?",
-    "仕事に戻る",
-    "おっぱい",
-    "Jimu ni modoru jikanda",
-    "Baiosenōshisu",
-    "Poruko poruko poruko",
-    "Bamosu jogā monoporio do Sonikku?",
-    "Kinshi no jikandesu",
-    "Afutā 8 aisukurīmu",
-    "Pika pika",
-]
