--- responses/Latin.py
@@ -1,34 +1,0 @@
-arr_latin = [
-    "Carpe diem quam minimum credula postero",
-    "Tua mater latior quam Rubicon est",
-    "Quid quid latine dictum sit, altum videtur",
-    "In vino veritas",
-    "Cave Canem",
-    "Dictum factum.",
-    "Dum vita est, spes est",
-    "Sapientia potentia est",
-    "Pax vobiscum",
-    "Suus 'tempus ad Duolingo lectionem",
-    "Accusare nemo se debet nisi coram Deo",
-    "Veni, vidi, vici",
-    "Sic parvis magna",
-    "Impetu vobiscum",
-    "Caecilius est in horto!",
-    "Per aspera ad astra",
-    "Audentes Fortuna iuvat",
-    "Alea iacta est",
-    "Et tu, Kiké?",
-    "Mea culpa",
-    "Cogito, ergo sum",
-    "Pedicabo ego vos et irrumabo",
-    "Loqui melius an mori?",
-    "Ave, Biocenose, morituri te salutant",
-    "Pro rata",
-    "Memento mori",
-    "Omnia vincit amor",
-    "Ceterum censeo Carthaginem delendam esse",
-    "Tempora mutantur et nos mutamur in illis",
-    "Forsan et haec olim meminisse iuvabit",
-    "Ignoramus et ignorabimus",
-    "A far l'amore comincia tu",
-    ]
